+++ src/rebuild_roms.rs
@@ -29,6 +29,14 @@ pub fn subcommand() -> Command {
                 .value_parser(PossibleValuesParser::new(MERGING_STRATEGIES)),
         )
         .arg(
+            Arg::new("SOLID")
+                .short('s')
+                .long("solid")
+                .help("Create solid 7z archives")
+                .required(false)
+                .action(ArgAction::SetTrue),
+        )
+        .arg(
             Arg::new("ALL")
                 .short('a')
                 .long("all")
@@ -74,7 +82,14 @@ pub async fn main(
             progress_bar.println("");
             continue;
         }
-        rebuild_system(connection, progress_bar, &system, merging).await?;
+        rebuild_system(
+            connection,
+            progress_bar,
+            &system,
+            merging,
+            matches.get_flag("SOLID"),
+        )
+        .await?;
         progress_bar.println("");
     }
 
@@ -86,6 +101,7 @@ async fn rebuild_system(
     progress_bar: &ProgressBar,
     system: &System,
     merging: Merging,
+    solid: bool,
 ) -> SimpleResult<()> {
     progress_bar.println(&format!("Processing \"{}\"", system.name));
 
@@ -95,7 +111,7 @@ async fn rebuild_system(
         && (merging == Merging::NonMerged || merging == Merging::FullNonMerged)
     {
         for game in games {
-            expand_game(connection, progress_bar, system, &game, merging).await?;
+            expand_game(connection, progress_bar, system, &game, merging, solid).await?;
         }
     } else if (system.merging == Merging::NonMerged as i64
         || system.merging == Merging::FullNonMerged as i64)
@@ -124,6 +140,7 @@ async fn expand_game(
     system: &System,
     game: &Game,
     merging: Merging,
+    solid: bool,
 ) -> SimpleResult<()> {
     progress_bar.println(&format!("Processing \"{}\"", game.name));
     let tmp_directory = get_tmp_directory(connection).await;
@@ -163,6 +180,7 @@ async fn expand_game(
             rom,
             &archive_romfile,
             tmp_directory,
+            solid,
         )
         .await?;
     }
@@ -223,6 +241,7 @@ async fn add_rom(
     rom: &Rom,
     archive_romfile: &Romfile,
     tmp_directory: &PathBuf,
+    solid: bool,
 ) -> SimpleResult<()> {
     match rom.romfile_id {
         Some(romfile_id) => {
@@ -241,6 +260,7 @@ async fn add_rom(
                         &archive_romfile.path,
                         &file_names,
                         tmp_directory,
+                        false,
                     )?;
                     remove_file(progress_bar, &tmp_directory.join(&rom.name), true).await?;
                 } else {
@@ -249,6 +269,7 @@ async fn add_rom(
                         &archive_romfile.path,
                         &file_names,
                         &Path::new(&romfile.path).parent().unwrap(),
+                        solid,
                     )?;
                 }
                 update_rom_romfile(transaction, rom.id, Some(archive_romfile.id)).await;
@@ -295,6 +316,7 @@ async fn add_rom(
                         &archive_romfile.path,
                         &[rom.name.as_str()],
                         tmp_directory,
+                        false,
                     )?;
                     remove_file(progress_bar, &tmp_directory.join(&rom.name), true).await?;
                 } else if existing_rom.name != rom.name {
@@ -310,6 +332,7 @@ async fn add_rom(
                         &archive_romfile.path,
                         &[rom.name.as_str()],
                         tmp_directory,
+                        solid,
                     )?;
                     remove_file(progress_bar, &tmp_directory.join(&rom.name), true).await?;
                 } else {
@@ -318,6 +341,7 @@ async fn add_rom(
                         &archive_romfile.path,
                         &file_names,
                         &Path::new(&existing_romfile.path).parent().unwrap(),
+                        solid,
                     )?;
                 }
                 update_rom_romfile(transaction, rom.id, Some(archive_romfile.id)).await;